fix(post-961): single hardening PR for P0–P3 residuals
Security / exports:
- MCP state_export passes localAdmin dual-intent (JOE-952)
- Daemon HTML CSP + nosniff/frame-ancestors on Mission Control
- Route tests for /storage/export and raw session messages
- API docs/OpenAPI dual-intent for storage export

Maintainability:
- Soft 1500 façades: scheduler, work-store, channel-commands; mcp ops extract
- Ratchet hard budgets (config/work/scheduler/work-store) + budget mcp/channel
- Local WorkspaceSessionPort progressive wiring via artifact handlers
- Wiki tool-router helpers peel; product knip in CI gateway/wiki

Ops honesty:
- Multi-writer readiness/doctor flag; HA docs strengthened
- Private-beta remains no-go; Dependabot #960 superseded note
- Cycle SCAN_ROOTS: document runtime-host residual cycle
