+++ products/gateway/src/cli/commands/health.ts
@@ -1,6 +1,7 @@
 import * as fs from 'node:fs'
 import * as path from 'node:path'
 import * as os from 'node:os'
+import { fileURLToPath } from 'node:url'
 import { z } from 'zod'
 import { getConfig } from '../../config.js'
 import { redactSecret } from '../../security.js'
@@ -142,6 +143,24 @@ export async function doctor() {
   console.log(`Local readiness catalog: ${blockedReadiness.length ? 'blocked' : attentionReadiness.length ? 'attention' : 'ready'} (${Object.entries(localReadiness.totals).map(([status, count]) => `${status}:${count}`).join(', ')})`)
   for (const entry of [...blockedReadiness, ...attentionReadiness].slice(0, 5)) {
     console.log(`- ${entry.id}: ${entry.statusCode} — ${entry.remediation || entry.summary}`)
+  }
+  // Multi-writer ownership posture (non-blocking): experimental multi-replica still fails open migrate hazards.
+  try {
+    const registryPath = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '../../../../../docs/development/distributed-ownership-proving-registry.json')
+    if (fs.existsSync(registryPath)) {
+      const registry = JSON.parse(fs.readFileSync(registryPath, 'utf8')) as { status?: string; openMigrateHazards?: string[] }
+      const open = Array.isArray(registry.openMigrateHazards) ? registry.openMigrateHazards : []
+      const registryStatus = typeof registry.status === 'string' ? registry.status : 'unknown'
+      if (open.length === 0 && registryStatus === 'ready') {
+        console.log(`Multi-writer ownership: ready (registry status=${registryStatus})`)
+      } else {
+        console.log(`Multi-writer ownership: single-daemon production only; experimental multi-replica still fails open migrate hazards (${open.join(', ') || 'none'}; registry status=${registryStatus})`)
+      }
+    } else {
+      console.log('Multi-writer ownership: proving registry not found (assume single-daemon production only)')
+    }
+  } catch {
+    console.log('Multi-writer ownership: could not load proving registry (assume single-daemon production only)')
   }
   console.log()
