+++ products/wiki/packages/mcp-server/src/tool-router.ts
@@ -2,14 +2,13 @@ import { configureGitRemote, diffVersions, getHistory, gitPull, gitPush, gitRemo
 import { createRun, runLocalJob } from "@openwiki/jobs";
 import { readCurrentIndexStoreGraph, readCurrentIndexStoreWorkspaceRegistry } from "@openwiki/index-store";
 import { readCurrentPostgresGraph, readCurrentPostgresRecord, readCurrentPostgresSource, readCurrentPostgresWorkspaceRegistry } from "@openwiki/postgres-runtime";
-import { assertAuthorized, inboxProcessRunInputId, pathAllowedByContextBounds, runJobAllowedFromMcp, runJobAuthorizationOperations, runJobAuthorizationSpec, runJobRequiresInboxItem } from "@openwiki/policy";
-import { listGraphEdges, loadRepository, readClaim, readDecision, readPage, readProposal, readProposalDetailWithOptions, readSource, readSourceContent, readWorkspaceRegistry, withRepositoryReadCache } from "@openwiki/repo";
+import { assertAuthorized, inboxProcessRunInputId, runJobAllowedFromMcp, runJobAuthorizationOperations, runJobAuthorizationSpec, runJobRequiresInboxItem } from "@openwiki/policy";
+import { listGraphEdges, loadRepository, readClaim, readDecision, readPage, readProposal, readSource, readSourceContent, readWorkspaceRegistry, withRepositoryReadCache } from "@openwiki/repo";
 import { searchWiki } from "@openwiki/search";
 import { publishStaticSite } from "@openwiki/static-export";
 import { validateRepository } from "@openwiki/validation";
 import { applyProposal, askWithCitations, commitChanges, dreamRunStatus, ignoreInboxItem, processInboxItem, readInboxWorkflow, redactThinkSearchExplainForPolicy, retryInboxItem, syncWorkspaceNow, thinkWithCitations, withWriteCoordination } from "@openwiki/workflows";
-import { redactOpenWikiRunRecord, synthesisTargetPath, writeOpenWikiLog, type ClaimRecord, type DecisionRecord, type PageRecord, type ProposalRecord, type RunRecord, type SearchRequest } from "@openwiki/core";
-import type { McpPolicyContext } from "./types.ts";
+import { synthesisTargetPath, writeOpenWikiLog, type ClaimRecord, type DecisionRecord, type PageRecord, type ProposalRecord, type SearchRequest } from "@openwiki/core";
 import {
   objectParams,
   stringParam,
@@ -44,7 +43,6 @@ import {
   toolAllowed,
   openWikiOperation,
   policyContextForMcp,
-  DEFAULT_LOCAL_MCP_ACTOR_ID,
   filterSearchResponseForMcp,
   assertMcpVisibleRecord,
   assertMcpInboxActionAuthorized,
@@ -83,6 +81,15 @@ import {
   traceClaimForMcp,
   fetchSourceFromMcp,
 } from "./tool-handlers.ts";
+import {
+  assertMcpPathsAuthorized,
+  dreamRunInputFromMcp,
+  mcpActorId,
+  readProposalDetailForMcp,
+  redactRunForMcp,
+  redactRunJobResult,
+  redactRunToolResponseForMcp,
+} from "./tool-router-helpers.ts";
 import { MCP_LIST_LIMIT_MAX, type McpServerOptions } from "./types.ts";
 
 const SOURCE_FETCH_AUTH_PATHS = ["sources/manifests", "sources/raw"] as const;
@@ -632,102 +639,4 @@ async function callToolInner(
     default:
       throw new Error(`Unsupported OpenWiki tool: ${name}`);
   }
-}
-
-function dreamRunInputFromMcp(args: Record<string, unknown>): Record<string, unknown> {
-  const phases = optionalStringArrayParam(args, "phases");
-  const limit = optionalNumberParam(args, "limit");
-  const timeoutMs = optionalNumberParam(args, "timeout_ms");
-  const dryRun = optionalBooleanParam(args, "dry_run");
-  const createProposals = optionalBooleanParam(args, "create_proposals");
-  return {
-    ...(phases === undefined ? {} : { phases }),
-    ...(limit === undefined ? {} : { limit }),
-    ...(timeoutMs === undefined ? {} : { timeout_ms: timeoutMs }),
-    ...(dryRun === undefined ? {} : { dry_run: dryRun }),
-    ...(createProposals === undefined ? {} : { create_proposals: createProposals }),
-    ...optionalStringObjectProperty(args, "provider", "provider"),
-    ...optionalStringObjectProperty(args, "schema_pack", "schema_pack"),
-  };
-}
-
-function mcpActorId(context: McpPolicyContext, args: Record<string, unknown>): string | undefined {
-  const requestedActorId = optionalStringParam(args, "actor_id");
-  if (context.actorId !== undefined && context.actorId !== DEFAULT_LOCAL_MCP_ACTOR_ID) {
-    return context.actorId;
-  }
-  if (context.role !== undefined || (context.principals !== undefined && context.principals.length > 0)) {
-    return "actor:user:local";
-  }
-  if (requestedActorId !== undefined) {
-    return requestedActorId;
-  }
-  if (context.actorId !== undefined) {
-    return context.actorId;
-  }
-  return undefined;
-}
-
-async function assertMcpPathsAuthorized(
-  root: string,
-  operation: Parameters<typeof assertMcpPathAuthorized>[1],
-  context: McpPolicyContext,
-  paths: readonly string[],
-): Promise<void> {
-  for (const repoPath of paths) {
-    await assertMcpPathAuthorized(root, operation, context, repoPath);
-  }
-}
-
-async function readProposalDetailForMcp(root: string, id: string, context: McpPolicyContext): Promise<unknown> {
-  const repo = await loadRepository(root);
-  const proposal = repo.proposals.find((candidate) => candidate.id === id || candidate.uri === id);
-  return readProposalDetailWithOptions(root, id, {
-    authorizePath(repoPath) {
-      if (proposal !== undefined && !proposalArtifactBelongsToProposal(proposal, repoPath)) {
-        throw new Error(`OpenWiki proposal artifact path is not bound to proposal ${proposal.id}: ${repoPath}`);
-      }
-      if (!pathAllowedByContextBounds(repo.policy, context, repoPath)) {
-        throw new Error(`OpenWiki policy bounds do not allow proposal artifact path ${repoPath}`);
-      }
-    },
-  });
-}
-
-function proposalArtifactBelongsToProposal(proposal: ProposalRecord, repoPath: string): boolean {
-  const stem = proposal.id.replace(/:/g, "_").replace(/-/g, "_");
-  return (
-    repoPath === `proposals/diffs/${stem}.diff` ||
-    repoPath === `proposals/reports/${stem}.json` ||
-    repoPath === `proposals/validation/${stem}.json` ||
-    repoPath.startsWith(`proposals/snapshots/${stem}/`)
-  );
-}
-
-function redactRunJobResult<T extends { run: RunRecord }>(result: T, context: McpPolicyContext): T {
-  return {
-    ...result,
-    run: redactRunForMcp(result.run, context),
-  };
-}
-
-function redactRunToolResponseForMcp(value: unknown, context: McpPolicyContext): unknown {
-  if (value !== null && typeof value === "object" && !Array.isArray(value) && "run" in value) {
-    const record = value as { run?: unknown };
-    if (isRunRecord(record.run)) {
-      return {
-        ...value,
-        run: redactRunForMcp(record.run, context),
-      };
-    }
-  }
-  return value;
-}
-
-function redactRunForMcp(run: RunRecord, context: McpPolicyContext): RunRecord {
-  return redactOpenWikiRunRecord(run, { includeSensitiveOperationalMetadata: context.role === "admin" || context.scopes.includes("wiki:admin") });
-}
-
-function isRunRecord(value: unknown): value is RunRecord {
-  return value !== null && typeof value === "object" && !Array.isArray(value) && (value as { type?: unknown }).type === "run";
 }
