+++ scripts/check-import-cycles.mjs
@@ -33,6 +33,16 @@ export const SCAN_ROOTS = [
   'packages/ui/src',
   // post-#959: widen ratchet toward runtime packages once proven cycle-free
   'packages/shared/src',
+  // post-#961 hardening: packages/runtime-host/src intentionally NOT added —
+  // known value-import cycle (not cycle-free at widen time):
+  //   runtime.ts
+  //     -> runtime-config-builder.ts
+  //       -> custom-agents-utils.ts
+  //         -> runtime-tools.ts
+  //           -> runtime.ts
+  // Break that cycle (e.g. extract pure tool-id helpers from runtime-tools, or
+  // move getClientForDirectory consumers off the runtime composition root)
+  // before ratcheting SCAN_ROOTS to include runtime-host.
 ]
 
 const SOURCE_EXTENSIONS = ['.ts', '.tsx']
